+++ src/jsMain/kotlin/org/snakeyaml/engine/v2/representer/platformRepresenter.js.kt
@@ -1,0 +1,7 @@
+package org.snakeyaml.engine.v2.representer
+
+import org.snakeyaml.engine.v2.api.DumpSettings
+
+actual fun Representer(settings: DumpSettings): Representer {
+    return CommonRepresenter(settings)
+}
